complaint-handling-system/backend/requirement.txt
flask==2.3.2
flask-cors==3.0.10
mysql-connector-python==8.0.33
python-dotenv==1.0.0
python-decouple==3.8